requirements.txt
Flask==0.12.2
pandas==0.21.0
pymongo==2.7.2
scikit-learn==0.19.1
matplotlib==2.1.1
mysql-connector==2.1.6
sklearn==0.0
scikit-learn==0.19.1
scipy==1.0.0
numpy==1.14.2
keras==2.2.4
tensorflow
ConfigParser